Test Code cleanup.
Mostly, this commit removes the usage of TestCases macros throughout the TD* code since these unit tests have been ported to our Tests project. Additionally, the setUp/tearDown methods of the classes in the Tests project were removed when they only called super. Finally, the init method in TDMultistreamWriter that was exposed in a previous update was wrapped around an ifdef DEBUG statement.