Fail CDT/TD_DatabaseManager when initWithDirectory path is file
The TDIsFileExistsError cannot distinguish between file or
directory error, so explicitly check for a file in the path
and not a directory in TD_DatabaseManager initWithDirectory:

TD_DatabaseManager initWithDirectory: returns nil if path is
a file or if some other kind of error occurs.

The CDTDatastoreManager fails to instantiate if TD_DatabaseManager
cannot be properly instantiated.

Add tests to make sure the CDTDatastoreManager fails to allocate
when initialized with a file that is in the way.

Add test to make sure the CDTDatastoreManager successfully
initializes when asked to initWithDirectory and that directory
already exists.