Using Delegate in example Project and fixing .pbxproj files
This fixes bugs reported in FB 28015 and 28017. The fix needs to
be done simultaneously (from what I can tell) because the example
project has a dependency on the Tests project (the Pods project
in the CDTDatastore workspace doesn't seem to be fully populated
until you run 'pod install' inside of Tests).

The example Project was verified to run in the simulator.