Maintain current attachments when updating documents.
This differs from CouchDB's API, which requires that all updates
include an _attachments dictionary. However, it feels more consistent
with the current attachments API, which is currently separate
from the main documents API.

It would be good to merge the documents and attachments API in future,
but a clean way hasn't presented itself as yet.