Add further attachment tests to ReplicationAcceptance
testReplicateManyLocalAttachments tests creating lots of local
attachments are successfully replicated.

testAddLocalReplicateDeleteLocalReplicate checks that deleting an
attachment from a local document that's already been replicated
remotely is replicated during a future replication.

testReplicateRemoteDocumentUpdate tests that updates to an
attachment are replicated correctly.

testReplicateRemoteDocumentDeleteLocalCheckReplicated tests that
deleting an attachment is replicated correctly.

testReplicateRemoteDocumentDeleteRemoteCheckReplicated tests that
deleting a remote attachment is replicated correctly.