Removes the _replicator database.
This commit removes the _replicator database that was used to control
and monitor replication processes. With this update, the TDReplicator
objects are created directly with TDReplicatorManager. The CDReplicator
objects are conceptually no longer wrappers around _replicator documents,
but instead are wrappers around the TDReplicator objects.

The CDTReplicator is initialized with a TDReplicatorManager object and
a NSDictionary that specifies the properties of the replication when
created by the CDTReplicatorFactory. There are significant changes to
CDTReplicator.m to make it reflect and monitor the state of the underlying
TDReplicator.

*References to the _replicator database are removed from TD_DatabaseManager.
*Adds new notification, TDReplicatorStartedNotification, to TDReplicator.
*Adds -createReplicatorWithProperties: and -startReplicator: to TDReplicatorManager.

By removing the _replicator database, however, there is now no support
to restart replications after the application returns from the background
or is restarted/crashed.

ReplicationAcceptance tests:
* Modifies a few tests to test CDTReplicator.changesTotal and changesProcessed properties.
The changesProcessed property, however, currently fails the test during a pull replication.
This issue will be addressed in future work.