Moves the TD_FilterBlock from TD_Database to TDPusher.
Previously, a TD_Database managed all filters associated with that
database in a local NSMutableDictionary. This has been removed and
now the TDPusher class holds a copy of the block which defines the
filter.

These changes are reflected in the CDT* level classes.

*filterId is removed from CDTPushReplication
*CDTReplicator is now initialized with an instance of an
CDTAbstractReplication class instead of its NSDictionary representation.
(This doesn't affect the API since this is a 'private' instantiation method.)
*CDTReplicator maintains a copy of a CDTFilterBlock
*CDTReplicatorFactory -oneWay was modified accordingly.