Fix bug where new create CRUD API didn't fire change notifications correctly
The notifications from TouchDB happened within a transaction for
the -createDocumentFromRevision:error: method, which probably meant
that any database calls happening in the notification callback didn't
see the changes the notification was in response to.

I therefore added a call to the notification handlers when we've
closed the transaction if it was successful.

I think there's still an issue with deleteDocumentById:error: right now,
but it needs further investigation.

RTC 52139