Adds error property in CDTReplicator for error information retrieval.
The error is translated from TD* level error type to CDT* error
here (the location for this transformation may change
in the future). If the TD* level error is not specifically
handled, the TD* error is used. As of this commit, only
TDReplicatorManagerErrorLocalDatabaseDeleted is transformed to
CDTReplicatorErrorLocalDatabaseDeleted.

Fixes small bug in CDTReplicatorErrors enum where the value for
CDTReplicatorErrorTDReplicatorWrongType was incorrecly set to 2.