Restores CDTDocumentRevision init method to be successful with nil body.
Previously one could init a CDTDocumentRevision with a nil body parameter.
That was changed in a recent update. While this hasn't caused any
known bugs up to now, there is a potential for something to be
overlooked. This update restores the original functionality since
there does not seem to be any known side-effects.