CDTReplicator stops its TDReplicator when it is deallocated.
Previously, the TDReplicator would continue working when the
CDTReplicator deallocated, which was inconsistent with the
expected behavior and would result in the CDTReplicator delegate
being uninformed of the state of replication.

The fire-and-forget is not yet fully in place, as now correctly
stated in the documentation.  This change ensures expected behavior -
replication stops when CDTReplicator is deallocated.

The developer must hold onto a strong reference to its CDTReplicator
in order for replication to continue.

Additionally, the internal TDReplicator.running BOOL is now set
immedately when TDReplicator -start is called and not after the
internal thread begins execution of the replication code. This
clarifies the state of the TDReplicator.