Cleans credentials from log messages.
This commit adds the TDCleanURLtoString function, which will
return a string representation of the URL with the
password value masked to '*****'. This works with most, but not
all NSURL objects. It does work with all expected NSURL objects
created by/for the CDTDatastore.

Tests were added for a number of URLs.