Ensures final remote checkpoint doc written.
Fixes bug where a replicator was not writing the last sequence
value to the remote checkpoint document because it's thread was
quiting before the request was made.

Prevents setting the TDReplicator error when the error is for
a local datastore being deleted. This was added so that the proper
error type is reported on a failure. In particular, when a local datastore
is deleted during a pull replication, it should report an error that the
datastore was deleted. Without this check in place, the error would be
overwritten by subsequent generic "Internal Error - 500". This error is
generated in the TDPuller callback block when it fails to write a local
document that it received from the remote.  (The TD_Database
returns a 590 - kTDStatusDBError, which is then converted into the
TDHTTP InternalError 500 by TDStatusToNSError function). Because
the replicator thread waits for all async tasks to complete before exiting
the thread, there can potentially be a number of requests for remote
documents still pending. All documents from the remote that are
pending a write to the local datastore will report this "500 DB Error". So,
deleting a local datastore is absolutely a fatal error during replication,
we prevent subsequent errors from overwriting the known fatal error.

This is part of this commit because it's necessary to pass all of the
tests in ReplicationAcceptance. One of those tests check the
error reported when a local datastore is deleted during replication. It failed
because the error no longer reported a deleted datastore but rather
the generic "500 DB Error" that was issued by the last callback block.