Removes the TDReplicator -clearDbRef method. This method seemed to be appropriate when all replications were executed on a main replication thread. This method served two purposes, previously.
First, it would assume a successful write to the remote checkpoint
document and write the last sequence value to the local checkpoint.
This is problematic if the write request to the remote checkpoint
document fails. In commit fa7c0ea, we've ensured the replicator
thread remains running until a callback block is called from the
last request to write to the remote checkpoint document. In that
callback block, we now write the local lastSequence value to the
checkpoint document if the remote write was successful. This ensures
they are synchronized. Writing the local value in -clearDbRef only
potentially confuses the situation and can result in a mismatch
between the local and remote checkpoint documents.

Additionally, -clearDbRef set the pointer to the TD_Database to nil.
It is now set to nil after the replicator's thread has exited its
runloop. This part of -clearDbRef is still necessary since it's one
of the mechanism to make each TDReplicator object only useable once.
The pointer to the TD_Database should not be set to nil until after
all asynchronous tasks have completed