Support arbitrary JSON seq nums for CouchDB 2.0
CouchDB 2.0 and Cloudant Local change sequence number format to be
an array rather than a number (as in Couch 1.0) or string (Cloudant)
which resulted in incorrect sequence numbers being stored and
used in the checkpoint document.

The fix is to support arbitrary JSON as the sequence number, so
we store that in the local docs database as { seq: <seq JSON> }
rather than as a raw string. The outer JSON is needed so that
the JSON serialiser will accept it, because old sequence numbers
are not valid JSON documents of themselves (being strings or
numbers).

Changing the seq format like this is a breaking changes, so we
have a migration which updates the existing checkpoint documents
and bumps the schema version to 100 to indicate that the seq
numbers are not usable in older versions.

Local databases store the seq as an auto-increment field in the
database, so we wrap that up into an object, an NSNumber in the
push code rather than wrapping as a string.