Create build script for CI Server
Heavily based on build.rb in sync-android. Handles the lifecycle of
docker containers, builds CDTDatastore for either iOS or OSX based on
the arguments given.

Defaults are:
- Couchdb 1.6
- OS X
- iPhone 4S (for iOS builds)
- Latest iOS Please enter the commit message for your changes. Lines starting

Added a buildplist bash script to modify the ReplicationSettings based on the environment
variables set for the shell it is running in during the build process.

BugzID: 45270