Add SQL translation support for text search
- Handle text clause separately from all other simple clauses since it
will be specific to a single text index (SQLite VIRTUAL table).  This
means that regardless of a compound clause's top level operator,
the text search clause will have its own separate SQL node in the query
node tree.
- Ensure that a query containing a text search will only execute if
all clauses of the query are satisfied by an index.  This includes the
text search clause of the query.