TDBlobStore uses a factory to create blob handles.
TDBlobStore is initialised now with a key provider; so it can create
instances that conform to protocols: CDTBlobReader, CDTBlobWriter &
CDTBlobMultipartWriter, to read/write encrypted or non-encrypted
attachments. Instead of adding inside TDBlobStore the code to
choose between the different types of CDTBlob*, we have defined the
class CDTBlobHandleFactory to do so.