Fix schema in ReplicationAcceptance project
Schema RA_Tests had to targets: iOS & OS X. But it failed if we tried
to run the tests for OS X. Also schema RA_Tests_OSX already run the
tests for OS X. So I deleted the second target in RA_Tests and modified
Rakefile to run the acceptance tests in each platform correctly.

Also, ReplicationAcceptance project had another schema to run an
app in OS X but this one did not even compile, so I removed it.