'CDTDatastore:getAllDocuments': Do not set a LIMIT
'CDTDatastore:getAllDocuments' uses
'TD_Database:getDocsWithIDs:options:' to get all the
documents in a datastore. It passes to
'TD_Database:getDocsWithIDs:options:' a struct called
query with the options, one of this options is the LIMIT.
This LIMIT was set to the number of documents in the
datastore, however
'TD_Database:getDocsWithIDs:options:' really works
with revisions, therefore this LIMIT is wrong. To fix
the problem we pass a LIMIT equal to the maximum
possible value: UINT_MAX.