Remove CDTMutableDocumentRevision
This commit removes the CDTMutableDocumentRevision class and fixes
users of the class to use the newly mutable CDTDocumentRevision class.

In this version, revision body and attachments are not mutable, and
users must make copies of them to change the content, then reassign
the new dictionary to the property. This adds boilerplate code where
changing revisions. This is fixed in a later commit.

In addition, conflict resolving tests are broken at this point as a
CDTDocumentRevision can't report whether it's been changed so the
resolving code must assume all resolved revision objects are new. This
creates unnecessary new revisions if the user's resolver choses an
existing revision as a resolution.