Make rev body and attachments mutable dictionaries
The aim of this patch is to make the body and attachments dictionaries
on CDTDocumentRevision objects directly mutatable, which simplifies
interacting with revision objects as the body and attachments arrays
don't need to be copied just to modify a subset of fields.

The problem which needs to be surmounted is that CDTDocumentRevision
still needs to be able to report whether its been changed since it
was retrieved from the database, and that's not possible with simple
foundation mutable collections.

This patch adds CDTChangedDictionary and CDTChangedArray. These
classes implement an observable wrapper around NSMutableDictionary and
NSMutableArray respectively. A delegate pattern is used for objects
of the class to inform interested parties of changes.

CDTDocumentRevision now creates a deeply nested copy of the body and
attachment data that it's initialised with as the Changed* containers.
The revision registers itself as delegate to the body and attachment
dictionaries so it can report to interested parties when the content
of the revision has changed, in particular conflict handling code.

Finally, the CDTDocumentRevision body and attachment dictionaries
become mutable dictionaries, allowing developers to easily modify
revision body and attachment dictionaries directly, saving boilerplate
code when changing revisions.