Use isFullRevision method to prevent data loss
Some operations, for example querying with a field filter, produce
document revision objects which are not complete representations
of the revision. In order to avoid data loss in applications, these
revisions should not be allowed to be saved by the datastore.

This patch introduces a method to CDTDocumentRevision, isFullRevision,
which allows these partial revision objects to be differentiated from
complete revisions. This allows the datastore to reject incomplete
revision objects.