Make Replication a one liner.
Added new methods to CDTDatastore+Replication to make replication easier to set up.
Added CDTDatastore+Replication to CloudantSync.h

Tests have been added to ReplicationAcceptance since they require a remote database.
The tests are hardcoded to use 10 documents as this is likely to complete within the
timeout for the expectation.

Resolves #268