Simplify podfile and update for new targets (#342)
Simplify podfile so it uses abstract targets rather than methods to define
common dependencies. Also make it possible to change the varients of pods
that are used depending on the envrionment, for example if encryption is
being used pods which enable encrpytion and testing of encrpyted datastores
is pulled in.