Added creds to UNIRest requests
The replication acceptance tests do some set up work by making requests
through UNIRest, which was not configured with any DB credentials. To
make the tests pass against a non admin-party host we need to add some
auth on the requests.
Added a basic auth header to all UNIRest requests if a user/pass was
set.